+++ common/src/main/java/com/infernalstudios/infernalexp/module/ModCreativeTabs.java
@@ -1,6 +1,7 @@
 package com.infernalstudios.infernalexp.module;
 
 import com.infernalstudios.infernalexp.IEConstants;
+import com.infernalstudios.infernalexp.compat.NetherExpCompat;
 import com.infernalstudios.infernalexp.registration.util.RegistrationProvider;
 import com.infernalstudios.infernalexp.registration.util.RegistryObject;
 import net.minecraft.core.registries.Registries;
@@ -69,6 +70,13 @@ public class ModCreativeTabs {
 
         entries.accept(ModBlocks.SHROOMLIGHT_TEAR.get());
         entries.accept(ModBlocks.HOLLOWLIGHT.get());
+
+        if (NetherExpCompat.SHROOMNIGHT_TEAR != null) {
+            entries.accept(NetherExpCompat.SHROOMNIGHT_TEAR.get());
+        }
+        if (NetherExpCompat.HOLLOWNIGHT != null) {
+            entries.accept(NetherExpCompat.HOLLOWNIGHT.get());
+        }
 
         entries.accept(ModBlocks.DULLTHORNS.get());
         entries.accept(ModBlocks.DULLTHORNS_BLOCK.get());